requirements.txt
django
requests
psycopg2-binary
djangorestframework
django-cors-headers